+++ CHANGELOG.md
@@ -11,13 +11,14 @@ NN XX 20NN
   - Emit error (not warning) for HTML4/XHTML1 strict doctypes
   - Fix bug that caused the checker to incorrectly treat `input[type]`
     values case-sensitively when doing particular checks
-  - Allow the `slot` attribute (for Shadow DOM interaction)
   - Further restrict script[type] and style[type]
+  - Allow the `slot` attribute (for Shadow DOM interaction)
   - Allow only one non-hidden `main` element per document
   - Allow only `html`, `body`, `div`, `form`, custom elements as `main` ancestor
   - Allow `role=none` where `role=presentation` is allowed
   - Allow `role=rowgroup` element to be contained in `role=table` element
   - Allow more ARIA states & properties for `role=menuitem`
+  - Disallow the `dropzone` attribute
 
 # 17.11.1
 07 October 2017